Configure warden when gem is included in the model
